+++ dummydata.js
@@ -1,0 +1,84 @@
+
+const listOfIdeas = [
+  {
+    id: 1,
+    title: 'Idea title',
+    description: 'Et amet ea dolor vero sed dolore ut et nonumy takimata, no rebum et erat erat dolore lorem, at sit kasd et sea dolor ipsum stet diam, magna diam sea takimata sea sed, no invidunt aliquyam sed dolores at ipsum voluptua, ut lorem et kasd at sed amet dolores consetetur.',
+    avatar_url: [require('/images/user_avatar.png')],
+    user_name: ['Shrilaskhmi', 'Koustov'],
+    status: 'InProgress'
+  },
+  {
+    id: 2,
+    title: 'Idea title2',
+    description: 'Et amet ea dolor vero sed dolore ut et nonumy takimata, no rebum et erat erat dolore lorem, at sit kasd et sea dolor ipsum stet diam, magna diam sea takimata sea sed, no invidunt aliquyam sed dolores at ipsum voluptua, ut lorem et kasd at sed amet dolores consetetur.',
+    avatar_url: [require('/images/user_avatar.png')],
+    user_name: ['Shrilaskhmi', 'Koustov'],
+    status: 'Submitted'
+  },
+  {
+    id: 3,
+    title: 'Idea title2',
+    description: 'Et amet ea dolor vero sed dolore ut et nonumy takimata, no rebum et erat erat dolore lorem, at sit kasd et sea dolor ipsum stet diam, magna diam sea takimata sea sed, no invidunt aliquyam sed dolores at ipsum voluptua, ut lorem et kasd at sed amet dolores consetetur.',
+    avatar_url: [require('/images/user_avatar.png')],
+    user_name: ['Shrilaskhmi', 'Koustov'],
+    status: 'Submitted'
+  },
+  {
+    id: 4,
+    title: 'Idea title2',
+    description: 'Et amet ea dolor vero sed dolore ut et nonumy takimata, no rebum et erat erat dolore lorem, at sit kasd et sea dolor ipsum stet diam, magna diam sea takimata sea sed, no invidunt aliquyam sed dolores at ipsum voluptua, ut lorem et kasd at sed amet dolores consetetur.',
+    avatar_url: [require('/images/user_avatar.png')],
+    user_name: ['Shrilaskhmi', 'Koustov'],
+    status: 'Not Started'
+  },
+  {
+    id: 5,
+    title: 'Idea title2',
+    description: 'Et amet ea dolor vero sed dolore ut et nonumy takimata, no rebum et erat erat dolore lorem, at sit kasd et sea dolor ipsum stet diam, magna diam sea takimata sea sed, no invidunt aliquyam sed dolores at ipsum voluptua, ut lorem et kasd at sed amet dolores consetetur.',
+    avatar_url: [require('/images/user_avatar.png')],
+    user_name: ['Shrilaskhmi', 'Koustov'],
+    status: 'Submitted'
+  },
+  {
+    id: 6,
+    title: 'Idea title2',
+    description: 'Et amet ea dolor vero sed dolore ut et nonumy takimata, no rebum et erat erat dolore lorem, at sit kasd et sea dolor ipsum stet diam, magna diam sea takimata sea sed, no invidunt aliquyam sed dolores at ipsum voluptua, ut lorem et kasd at sed amet dolores consetetur.',
+    avatar_url: [require('/images/user_avatar.png')],
+    user_name: ['Shrilaskhmi', 'Koustov'],
+    status: 'Submitted'
+  },
+  {
+    id: 7,
+    title: 'Idea title2',
+    description: 'Et amet ea dolor vero sed dolore ut et nonumy takimata, no rebum et erat erat dolore lorem, at sit kasd et sea dolor ipsum stet diam, magna diam sea takimata sea sed, no invidunt aliquyam sed dolores at ipsum voluptua, ut lorem et kasd at sed amet dolores consetetur.',
+    avatar_url: [require('/images/user_avatar.png')],
+    user_name: ['Shrilaskhmi', 'Koustov'],
+    status: 'Submitted'
+  },
+  {
+    id: 8,
+    title: 'Idea title2',
+    description: 'Et amet ea dolor vero sed dolore ut et nonumy takimata, no rebum et erat erat dolore lorem, at sit kasd et sea dolor ipsum stet diam, magna diam sea takimata sea sed, no invidunt aliquyam sed dolores at ipsum voluptua, ut lorem et kasd at sed amet dolores consetetur.',
+    avatar_url: [require('/images/user_avatar.png')],
+    user_name: ['Shrilaskhmi', 'Koustov'],
+    status: 'Submitted'
+  },
+  {
+    id: 9,
+    title: 'Idea title2',
+    description: 'Et amet ea dolor vero sed dolore ut et nonumy takimata, no rebum et erat erat dolore lorem, at sit kasd et sea dolor ipsum stet diam, magna diam sea takimata sea sed, no invidunt aliquyam sed dolores at ipsum voluptua, ut lorem et kasd at sed amet dolores consetetur.',
+    avatar_url: [require('/images/user_avatar.png')],
+    user_name: ['Shrilaskhmi', 'Koustov'],
+    status: 'Submitted'
+  },
+  {
+    id: 10,
+    title: 'Idea title2',
+    description: 'Et amet ea dolor vero sed dolore ut et nonumy takimata, no rebum et erat erat dolore lorem, at sit kasd et sea dolor ipsum stet diam, magna diam sea takimata sea sed, no invidunt aliquyam sed dolores at ipsum voluptua, ut lorem et kasd at sed amet dolores consetetur.',
+    avatar_url: [require('/images/user_avatar.png'), require('/images/user_avatar.png')],
+    user_name: ['Shrilaskhmi', 'Koustov'],
+    status: 'Submitted'
+  }
+];
+export { listOfIdeas };
