+++ package.json
@@ -11,6 +11,7 @@
   "dependencies": {
     "@nhost/nextjs": "^1.7.4",
     "@nhost/react": "^0.12.4",
+    "axios": "^0.27.2",
     "graphql": "^16.6.0",
     "next": "12.2.5",
     "react": "18.2.0",
