+++ styles/idea.module.css
@@ -1,7 +1,7 @@
 .container{
   background-color: #ffffff;
+  width: 100%;
   height: 100vh;
-  width: 100%; 
   padding: 1rem;
 }
